fix: refactor footerWithFab component
